fixed display of exp double converted to int truncates instead, doubles should be printed as %.0f if you want to trim the decimal

+++ code/being.cc
@@ -1372,7 +1372,7 @@ sstring TBeing::displayExp() const
   if (getExp() < 100)
     sprintf(buf, "%.3f", getExp());
   else
-    sprintf(buf, "%d", (int) getExp());
+    sprintf(buf, "%.0f", getExp());
 #else
   if (getExp() < 100)
     sprintf(buf, "%.3f", getExp());
